US26: Refactor: rm merchnt status and update->edit
